+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Fix title margin rules
 * Add JS function to convert image to base64 code
 * Fix download all button action
+* Label sizes
+* Add input boder style to customizer
 
 ## 1.65.29 | 22.09.2022 <💎 IMPROVE & 📦 NEW: customizer>
 * h1 - h6 title to customizer tag styling